[REF] VER 01

+++ src/main/java/com/controle/repository/ControleContaRepository.java
@@ -3,6 +3,6 @@ package com.controle.repository;
 import com.controle.model.ControleConta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ControleContaRepository extends JpaRepository<ControleConta,Long>{
+public interface ControleContaRepository extends JpaRepository<ControleConta, Integer> {
 
 }
